refactor: move casos de uso para camada de aplicacao
