+++ src/test/java/br/edu/ifsp/foodflow/app/domain/order/RemoveItemFromOrderUseCaseTest.java
@@ -4,7 +4,7 @@ import br.edu.ifsp.foodflow.app.domain.addOn.AddOn;
 import br.edu.ifsp.foodflow.app.domain.menuItem.MenuItem;
 import br.edu.ifsp.foodflow.app.domain.order.dto.OrderResponse;
 import br.edu.ifsp.foodflow.app.domain.order.dto.RemoveItemFromOrderRequest;
-import br.edu.ifsp.foodflow.app.domain.order.useCases.RemoveItemFromOrderUseCase;
+import br.edu.ifsp.foodflow.app.application.useCases.order.RemoveItemFromOrderUseCase;
 import br.edu.ifsp.foodflow.app.domain.orderItem.OrderItem;
 import br.edu.ifsp.foodflow.app.domain.orderItem.OrderItemRepository;
 import br.edu.ifsp.foodflow.app.domain.orderItem.OrderItemStatus;
